azfunctions/requirements.txt
azure-core==1.34.0
azure-functions==1.23.0
azure-servicebus==7.14.2
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
isodate==0.7.2
json-logic==0.6.3
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
